split tests into functional and integration
